pyproject.toml
[project]
name = "python-project-template"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.2",
    "langchain-openai>=1.0.1",
    "openai>=2.6.0",
    "python-dotenv>=1.1.1",
]

[tool.mypy]
strict = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "ANN", "TRY", "EM", "FA", "SIM108", "S101", "RET505", "TCH002"]
